Save to log passed log file
When you save, it will write into the file of the log you passed in with --log, but use the regular file creation if it wasn't passed in. If you reset or rescramble the puzzle and save, it will use a newly generated file when you so it won't overwrite the log file from the previous solve.